% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_calib}
\alias{control_calib}
\title{control parameters}
\usage{
control_calib(
  interpolation = c("logit", "linear"),
  logit_const = -1000,
  sparse = FALSE
)
}
\arguments{
\item{interpolation}{type of interpolation: \code{logit} or \code{linear}}

\item{logit_const}{constant for \code{logit} interpolation}

\item{sparse}{whether to use sparse matrices via \code{Matrix} package}
}
\value{
a list with parameters
}
\description{
\code{control_calib} is function that contains control parameters for \code{joint_calib_create_matrix}
}
\author{
Maciej Beręsewicz
}
