% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_calib_att.R
\name{joint_calib_att}
\alias{joint_calib_att}
\title{Function for the balancing control to treatment groups using \code{joint_calib}}
\usage{
joint_calib_att(
  formula_means = NULL,
  formula_quantiles = NULL,
  treatment = NULL,
  data,
  probs = c(0.25, 0.5, 0.75),
  ...
)
}
\arguments{
\item{formula_means}{a formula with variables to be balanced at means,}

\item{formula_quantiles}{a formula with variables to be balanced at quantiles,}

\item{treatment}{a formula with treatment indicator}

\item{data}{a data.frame with variables,}

\item{probs}{a vector or a named list with quantiles to be balanced,}

\item{...}{other parameters passed to \code{joint_calib} function.}
}
\value{
Returns a list with containing:\cr
\itemize{
\item{\code{g} -- g-weight that sums up to sample size,}
\item{\code{Xs} -- matrix used for calibration (i.e. Intercept, X and X_q transformed for calibration of quantiles),}
\item{\code{totals} -- a vector of totals (i.e. \code{N}, \code{pop_totals} and \code{pop_quantiles}),}
\item{\code{method} -- selected method,}
\item{\code{backend} -- selected backend.}
}
}
\description{
\code{joint_calib} allows joint calibration of totals and quantiles. It provides a user-friendly interface that includes the specification of variables in formula notation, a vector of population totals, a list of quantiles, and a variety of backends and methods.
}
\examples{
# generate data as in hbal package
set.seed(123)
N <- 1500
X1 <- rnorm(N)
X2 <- rnorm(N)
X3 <- rbinom(N, size = 1, prob = .5)
X1X3 <- X1*X3
D_star <- 0.5*X1 + 0.3*X2 + 0.2*X1*X2 - 0.5*X1*X3 - 1
D <- ifelse(D_star > rnorm(N), 1, 0) # Treatment indicator
y <- 0.5*D + X1 + X2 + X2*X3 + rnorm(N) # Outcome
dat <- data.frame(D = D, X1 = X1, X2 = X2, X3 = X3, X1X3=X1X3, Y = y)
head(dat)

results <- joint_calib_att(
formula_means = ~ X1 + X2 + X3,
formula_quantiles = ~ X1 + X2,
treatment = ~ D,
data = dat,
method = "raking"
)

results

## interaction for means
results2 <- joint_calib_att(
formula_means = ~ X1 + X2 + X3 + X1*X3,
formula_quantiles = ~ X1 + X2,
treatment = ~ D,
data = dat,
method = "raking"
)

results2

}
\references{
reference
}
\author{
Maciej Beręsewicz
}
