% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.jointCalib}
\alias{print.jointCalib}
\title{Print method for jointCalib class}
\usage{
\method{print}{jointCalib}(x, ...)
}
\arguments{
\item{x}{object of jointCalib class}

\item{...}{Additional optional arguments}
}
\description{
Print method for jointCalib class
}
