% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_el.R
\name{calib_el}
\alias{calib_el}
\title{Weights calibration using empirical likelihood function}
\usage{
calib_el(X, d, totals, ...)
}
\arguments{
\item{X}{matrix of variables for calibration of quantiles and totals}

\item{d}{initial d-weights for calibration (e.g. design-weights)}

\item{totals}{vector of totals (where 1 element is population size)}

\item{...}{arguments passed to \code{stats::constrOptim}}
}
\value{
Returns a vector of empirical likelihood weights that sums to N
}
\description{
\code{calib_el} performs calibration using empirical likelihood method. The function is taken from the supplementary materials from the publication Zhang, Han and Wu (2022)
}
\references{
Zhang, S., Han, P., and Wu, C. (2023) Calibration Techniques Encompassing Survey Sampling, Missing Data Analysis and Causal Inference. International Statistical Review, 91: 165–192. https://doi.org/10.1111/insr.12518.
}
\author{
Maciej Beręsewicz based on Zhang, Han and Wu (2022)

\loadmathjax
}
