% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_el.R
\name{calib_el}
\alias{calib_el}
\title{Weights calibration using empirical likelihood function}
\usage{
calib_el(X, d, totals, maxit, tol)
}
\arguments{
\item{X}{matrix of variables for calibration of quantiles and totals}

\item{d}{initial d-weights for calibration (e.g. design-weights)}

\item{totals}{vector of totals (where 1 element is population size)}

\item{maxit}{a numeric value giving the maximum number of iterations}

\item{tol}{the desired accuracy for the iterative procedure}
}
\value{
Returns a vector of empirical likelihood weights that sums to N
}
\description{
\code{calib_el} performs calibration using empirical likelihood method. The function is taken from Wu (2005)
}
\references{
Wu, C. (2005). Algorithms and R codes for the pseudo empirical likelihood method in survey sampling. Survey Methodology, 31(2), 239..
}
\author{
Maciej Beręsewicz based on Zhang, Han and Wu (2022)

\loadmathjax
}
