% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_calib}
\alias{control_calib}
\title{control parameters}
\usage{
control_calib(
  interpolation = c("logit", "linear"),
  logit_const = -1000,
  survey_sparse = FALSE,
  ebal_constraint_tolerance = 1,
  ebal_print_level = 0
)
}
\arguments{
\item{interpolation}{type of interpolation: \code{logit} or \code{linear},}

\item{logit_const}{constant for \code{logit} interpolation,}

\item{survey_sparse}{whether to use sparse matrices via \code{Matrix} package in \code{\link[survey:calibrate]{survey::grake()}} (currently not supported),}

\item{ebal_constraint_tolerance}{This is the tolerance level used by ebalance to decide if the moments in the reweighted data are equal to the target moments (see \code{\link[ebal:ebalance]{ebal::ebalance()}}),}

\item{ebal_print_level}{Controls the level of printing: 0 (normal printing), 2 (detailed), and 3 (very detailed) (see \code{\link[ebal:ebalance]{ebal::ebalance()}}).}
}
\value{
a list with parameters
}
\description{
\code{control_calib} is function that contains control parameters for \code{joint_calib_create_matrix}
}
\author{
Maciej Beręsewicz
}
