% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_calib_att.R
\name{joint_calib_att}
\alias{joint_calib_att}
\title{Function for balancing the control and treatment group using \code{joint_calib}}
\usage{
joint_calib_att(
  formula_means = NULL,
  formula_quantiles = NULL,
  treatment = NULL,
  data,
  probs = c(0.25, 0.5, 0.75),
  ...
)
}
\arguments{
\item{formula_means}{a formula with variables to be balanced at means,}

\item{formula_quantiles}{a formula with variables to be balanced at quantiles,}

\item{treatment}{a formula with treatment indicator}

\item{data}{a data.frame with variables,}

\item{probs}{a vector or a named list with quantiles to be balanced (default is \code{c(0.25, 0.5, 0.75)}),}

\item{...}{other parameters passed to \code{joint_calib} function.}
}
\value{
Returns a list with containing:\cr
\itemize{
\item{\code{g} -- g-weight that sums up to treatment group size,}
\item{\code{Xs} -- matrix used for calibration (i.e. Intercept, X and X_q transformed for calibration of quantiles),}
\item{\code{totals} -- a vector of totals (i.e. \code{N}, \code{pop_totals} and \code{pop_quantiles}),}
\item{\code{method} -- selected method,}
\item{\code{backend} -- selected backend.}
}
}
\description{
\code{joint_calib_att} allows quantile or mean and quantile balancing of control and treatment groups. It provides a user-friendly interface to specify the variables and quantiles to be balanced.
}
\examples{
# generate data as in hbal package
set.seed(123)
N <- 1500
X1 <- rnorm(N)
X2 <- rnorm(N)
X3 <- rbinom(N, size = 1, prob = .5)
X1X3 <- X1*X3
D_star <- 0.5*X1 + 0.3*X2 + 0.2*X1*X2 - 0.5*X1*X3 - 1
D <- ifelse(D_star > rnorm(N), 1, 0) # Treatment indicator
y <- 0.5*D + X1 + X2 + X2*X3 + rnorm(N) # Outcome
dat <- data.frame(D = D, X1 = X1, X2 = X2, X3 = X3, X1X3=X1X3, Y = y)
head(dat)

results <- joint_calib_att(
formula_means = ~ X1 + X2 + X3,
formula_quantiles = ~ X1 + X2,
treatment = ~ D,
data = dat,
method = "raking"
)

results

## interaction for means
results2 <- joint_calib_att(
formula_means = ~ X1 + X2 + X3 + X1*X3,
formula_quantiles = ~ X1 + X2,
treatment = ~ D,
data = dat,
method = "raking"
)

results2

## probs as a named list with varying probs
results3 <- joint_calib_att(
formula_means = ~ X1 + X2 + X3 + X1*X3,
formula_quantiles = ~ X1 + X2,
treatment = ~ D,
data = dat,
method = "raking",
probs = list(X1 = 0.5, X2 = c(0.25, 0.75))
)

results3

## probs for interactions
results4 <- joint_calib_att(
formula_means = ~ X1 + X2 + X3,
formula_quantiles = ~ X1 + X2 + X1:X3,
treatment = ~ D,
data = dat,
probs = list(X1=0.5, X2 = c(0.25, 0.5), `X1:X3` = 0.75),
method = "raking"
)

results4

}
\references{
Greifer N (2023). WeightIt: Weighting for Covariate Balance in Observational Studies.
R package version 0.14.2, \url{https://CRAN.R-project.org/package=WeightIt}.

Greifer N (2023). cobalt: Covariate Balance Tables and Plots.
R package version 4.5.1, \url{https://CRAN.R-project.org/package=cobalt}.

Ho, D., Imai, K., King, G., & Stuart, E. A. (2011).
MatchIt: Nonparametric Preprocessing for Parametric Causal Inference.
Journal of Statistical Software, 42(8), 1–28. \url{https://doi.org/10.18637/jss.v042.i08}

Xu, Y., & Yang, E. (2023). Hierarchically Regularized Entropy Balancing.
Political Analysis, 31(3), 457-464. \url{https://doi.org/10.1017/pan.2022.12}
}
\author{
Maciej Beręsewicz
}
